[FEATURE] Changed AppName to "ListAssist" / Added Spam Protection to Buttons

+++ lib/widgets/shoppinglist/shopping_list_detail.dart
@@ -29,6 +29,9 @@ class _ShoppingListDetail extends State<ShoppingListDetail> {
   String uid = "";
   bool useCache = false;
 
+  //Spamschutz z.B. beim Löschen
+  bool _buttonsDisabled = false;
+
   Timer _debounce;
   int _debounceTime = 1500;
 
@@ -111,52 +114,61 @@ class _ShoppingListDetail extends State<ShoppingListDetail> {
       ),
       floatingActionButton: Stack(
         children: <Widget>[
-          Padding(
-            padding: EdgeInsets.only(right: 80.0),
-            child: Align(
-              alignment: Alignment.bottomRight,
-              child: FloatingActionButton(
-                child: Icon(Icons.add),
-                backgroundColor: Colors.green,
-                onPressed: () {
-                  Navigator.push(context, MaterialPageRoute(builder: (context) => SearchItemsView(lists.elementAt(widget.index).id)));
-                },
-              ),
+          Align(
+            alignment: Alignment.bottomRight,
+            child: FloatingActionButton(
+              child: Icon(Icons.add),
+              backgroundColor: Colors.green,
+              onPressed: () {
+                Navigator.push(context, MaterialPageRoute(builder: (context) => SearchItemsView(lists.elementAt(widget.index).id)));
+              },
             ),
           ),
-          SpeedDial(
-            animatedIcon: AnimatedIcons.menu_close,
-            animatedIconTheme: IconThemeData(size: 22.0),
-            closeManually: false,
-            curve: Curves.easeIn,
-            overlayOpacity: 0.35,
-            backgroundColor: Theme.of(context).primaryColor,
-            elevation: 8.0,
-            shape: CircleBorder(),
-            children: [
-              SpeedDialChild(
-                  child: Icon(Icons.check),
-                  backgroundColor: Colors.green,
+          Padding(
+            padding: EdgeInsets.only(bottom : 75.0),
+            child: SpeedDial(
+              animatedIcon: AnimatedIcons.menu_close,
+              animatedIconTheme: IconThemeData(size: 22.0),
+              closeManually: false,
+              curve: Curves.easeIn,
+              overlayOpacity: 0.35,
+              backgroundColor: Theme.of(context).primaryColor,
+              elevation: 8.0,
+              shape: CircleBorder(),
+              children: [
+                SpeedDialChild(
+                    child: Icon(Icons.check),
+                    backgroundColor: Colors.green,
+                    labelBackgroundColor: Theme.of(context).brightness == Brightness.dark ? Theme.of(context).primaryColor : Colors.white,
+                    label: "Complete",
+                    labelStyle: TextStyle(fontSize: 18.0, color: Theme.of(context).brightness == Brightness.dark ? Colors.white : Colors.black),
+                    onTap: _showCompleteDialog),
+                SpeedDialChild(
+                    child: Icon(Icons.delete),
+                    backgroundColor: Colors.red,
+                    labelBackgroundColor: Theme.of(context).brightness == Brightness.dark ? Theme.of(context).primaryColor : Colors.white,
+                    label: "Delete",
+                    labelStyle: TextStyle(fontSize: 18.0, color: Theme.of(context).brightness == Brightness.dark ? Colors.white : Colors.black),
+                    onTap: _showDeleteDialog),
+                SpeedDialChild(
+                  child: Icon(Icons.camera),
+                  backgroundColor: Colors.blue,
+                  label: "Image Check",
                   labelBackgroundColor: Theme.of(context).brightness == Brightness.dark ? Theme.of(context).primaryColor : Colors.white,
-                  label: "Complete",
                   labelStyle: TextStyle(fontSize: 18.0, color: Theme.of(context).brightness == Brightness.dark ? Colors.white : Colors.black),
-                  onTap: _showCompleteDialog),
-              SpeedDialChild(
-                  child: Icon(Icons.delete),
-                  backgroundColor: Colors.red,
-                  labelBackgroundColor: Theme.of(context).brightness == Brightness.dark ? Theme.of(context).primaryColor : Colors.white,
-                  label: "Delete",
-                  labelStyle: TextStyle(fontSize: 18.0, color: Theme.of(context).brightness == Brightness.dark ? Colors.white : Colors.black),
-                  onTap: _showDeleteDialog),
-              SpeedDialChild(
-                child: Icon(Icons.camera),
-                backgroundColor: Colors.blue,
-                label: "Image Check",
-                labelBackgroundColor: Theme.of(context).brightness == Brightness.dark ? Theme.of(context).primaryColor : Colors.white,
-                labelStyle: TextStyle(fontSize: 18.0, color: Theme.of(context).brightness == Brightness.dark ? Colors.white : Colors.black),
-                onTap: () => InfoOverlay.showSourceSelectionSheet(context, callback: _startCameraScanner, arg: widget.index),
-              )
-            ],
+                  onTap: () async{
+                    bool connected = await connectivityService.testInternetConnection();
+                    if (!connected) {
+                      //I am NOT connected to the Internet
+                      InfoOverlay.showErrorSnackBar("Kein Internetzugriff");
+                      _buttonsDisabled = false;
+                    } else {
+                      InfoOverlay.showSourceSelectionSheet(context, callback: _startCameraScanner, arg: widget.index);
+                    }
+                  },
+                )
+              ],
+            ),
           ),
         ],
       ),
@@ -205,22 +217,34 @@ class _ShoppingListDetail extends State<ShoppingListDetail> {
             ),
             FlatButton(
               child: Text("Abschließen"),
-              onPressed: () {
-                useCache = true;
-                list = ShoppingList(
-                  id: list.id,
-                  created: list.created,
-                  name: list.name,
-                  items: list.items,
-                );
-                databaseService.completeList(uid, list).catchError((_) {
-                  InfoOverlay.showErrorSnackBar("Fehler beim Abschließen der Einkaufsliste");
-                  useCache = false;
-                }).then((_) {
-                  InfoOverlay.showInfoSnackBar("Einkaufsliste ${list.name} abgeschlossen");
-                  Navigator.of(context).pop();
-                  Navigator.of(this.context).pop();
-                });
+              onPressed: () async {
+                if (!_buttonsDisabled) {
+                  bool connected = await connectivityService.testInternetConnection();
+                  if (!connected) {
+                    //I am NOT connected to the Internet
+                    InfoOverlay.showErrorSnackBar("Kein Internetzugriff");
+                    _buttonsDisabled = false;
+                  } else {
+                    //I am connected to the Internet
+                    _buttonsDisabled = true;
+                    useCache = true;
+                    list = ShoppingList(
+                      id: list.id,
+                      created: list.created,
+                      name: list.name,
+                      items: list.items,
+                    );
+                    databaseService.completeList(uid, list).catchError((_) {
+                      InfoOverlay.showErrorSnackBar("Fehler beim Abschließen der Einkaufsliste");
+                      useCache = false;
+                      _buttonsDisabled = false;
+                    }).then((_) {
+                      InfoOverlay.showInfoSnackBar("Einkaufsliste ${list.name} abgeschlossen");
+                      Navigator.of(context).pop();
+                      Navigator.of(this.context).pop();
+                    });
+                  }
+                }
               },
             ),
           ],
@@ -262,29 +286,34 @@ class _ShoppingListDetail extends State<ShoppingListDetail> {
             ),
             FlatButton(
               child: Text("Löschen"),
-              onPressed: () async{
-                useCache = true;
-                list = ShoppingList(
-                  id: list.id,
-                  created: list.created,
-                  name: list.name,
-                  items: list.items,
-                );
+              onPressed: () async {
+                if (!_buttonsDisabled) {
+                  _buttonsDisabled = true;
+                  useCache = true;
+                  list = ShoppingList(
+                    id: list.id,
+                    created: list.created,
+                    name: list.name,
+                    items: list.items,
+                  );
 
-                bool connected = await connectivityService.testInternetConnection();
-                if (!connected) {
-                  //I am NOT connected to the Internet
-                  InfoOverlay.showErrorSnackBar("Kein Internetzugriff");
-                } else {
-                  //I am connected to the Internet
-                  databaseService.deleteList(uid, list.id).catchError((_) {
-                    InfoOverlay.showErrorSnackBar("Fehler beim Löschen der Einkaufsliste");
-                    useCache = false;
-                  }).then((_) {
-                    InfoOverlay.showInfoSnackBar("Einkaufsliste ${list.name} gelöscht");
-                    Navigator.of(context).pop();
-                    Navigator.of(this.context).pop();
-                  });
+                  bool connected = await connectivityService.testInternetConnection();
+                  if (!connected) {
+                    //I am NOT connected to the Internet
+                    InfoOverlay.showErrorSnackBar("Kein Internetzugriff");
+                    _buttonsDisabled = false;
+                  } else {
+                    //I am connected to the Internet
+                    databaseService.deleteList(uid, list.id).catchError((_) {
+                      InfoOverlay.showErrorSnackBar("Fehler beim Löschen der Einkaufsliste");
+                      useCache = false;
+                      _buttonsDisabled = false;
+                    }).then((_) {
+                      InfoOverlay.showInfoSnackBar("Einkaufsliste ${list.name} gelöscht");
+                      Navigator.of(context).pop();
+                      Navigator.of(this.context).pop();
+                    });
+                  }
                 }
               },
             ),
